refactor: UserDetails 관련 객체 응집
